LayoutTests/fast/text/font-face-set-remove-safely.html

<script>
if (window.testRunner)
    testRunner.dumpAsText();

function test() {
   console.log("Retrieving fonts...");
   var fonts = document.fonts;
   console.log("Clearing...");
   fonts.clear();
   console.log("Done.");
}

</script>
<body onload=test()>

<svg>

<font>

<font-face font-family="Palatino, Menlo, Verdana, serif, monospace">
