LayoutTests/fast/text/font-face-set-remove-safely-expected.txt

CONSOLE MESSAGE: line 6: Retrieving fonts...
CONSOLE MESSAGE: line 8: Clearing...
CONSOLE MESSAGE: line 10: Done.
